+++ src/main/java/com/github/zipcodewilmington/casino/games/craps/CrapsPlayer.java
@@ -2,8 +2,9 @@ package com.github.zipcodewilmington.casino.games.craps;
 
 import com.github.zipcodewilmington.casino.CasinoAccount;
 import com.github.zipcodewilmington.casino.Player;
+import com.github.zipcodewilmington.casino.WagingPlayer;
 
-public class CrapsPlayer extends Player {
+public class CrapsPlayer extends WagingPlayer {
     @Override
     public CasinoAccount getCasinoAccount() {
         return null;
@@ -13,4 +14,6 @@ public class CrapsPlayer extends Player {
     public <SomeReturnType> SomeReturnType play() {
         return null;
     }
+
+
 }
