+++ src/main/java/com/github/zipcodewilmington/casino/games/klondike/KlondikePlayer.java
@@ -2,8 +2,9 @@ package com.github.zipcodewilmington.casino.games.klondike;
 
 import com.github.zipcodewilmington.casino.CasinoAccount;
 import com.github.zipcodewilmington.casino.Player;
+import com.github.zipcodewilmington.casino.WagingPlayer;
 
-public class KlondikePlayer extends Player {
+public class KlondikePlayer extends WagingPlayer {
     @Override
     public CasinoAccount getCasinoAccount() {
         return this.casinoAccount;
@@ -14,8 +15,5 @@ public class KlondikePlayer extends Player {
         return null;
     }
 
-    @Override
-    public void run() {
 
-    }
 }
