+++ src/main/java/com/github/zipcodewilmington/casino/games/numberguess/NumberGuessPlayer.java
@@ -2,11 +2,12 @@ package com.github.zipcodewilmington.casino.games.numberguess;
 
 import com.github.zipcodewilmington.casino.CasinoAccount;
 import com.github.zipcodewilmington.casino.Player;
+import com.github.zipcodewilmington.casino.WagingPlayer;
 
 /**
  * Created by leon on 7/21/2020.
  */
-public class NumberGuessPlayer extends Player {
+public class NumberGuessPlayer extends WagingPlayer {
     @Override
     public CasinoAccount getCasinoAccount() {
 
@@ -17,4 +18,6 @@ public class NumberGuessPlayer extends Player {
     public <SomeReturnType> SomeReturnType play() {
         return null;
     }
+
+
 }
