+++ src/main/java/com/github/zipcodewilmington/casino/games/war/WarPlayer.java
@@ -14,8 +14,5 @@ public class WarPlayer extends Player {
         return null;
     }
 
-    @Override
-    public void run() {
 
-    }
 }
